Move DisplayPatterns to main project
