+++ LegacySignController/src/ArduinoHeader.h
@@ -1,0 +1,10 @@
+#ifndef ARDUINOHEADER_H
+#define ARDUINOHEADER_H
+
+#ifdef PIO_UNIT_TESTING
+#include <ArduinoFake.h>
+#else
+#include <Arduino.h>
+#endif
+
+#endif
