+++ LegacySignController/src/NeoPixelDisplay.h
@@ -6,7 +6,7 @@
 #include <Adafruit_NeoPixel.h>
 #include <PixelMap.h>
 #include <DisplayConfiguration.h>
-#include <DisplayPattern.h>
+#include "Patterns\DisplayPatterns\DisplayPattern.h"
 
 class NeoPixelDisplay
 {
