+++ LegacySignController/src/Patterns/DisplayPatterns/CenterOutDisplayPattern.cpp
