+++ LegacySignController/src/Patterns/DisplayPatterns/CenterOutDisplayPattern.h
