+++ LegacySignController/src/Patterns/DisplayPatterns/DisplayPattern.cpp
