+++ LegacySignController/src/Patterns/DisplayPatterns/DisplayPattern.h
