+++ LegacySignController/src/Patterns/DisplayPatterns/FireDisplayPattern.cpp
