+++ LegacySignController/src/Patterns/DisplayPatterns/FireDisplayPattern.h
