+++ LegacySignController/src/Patterns/DisplayPatterns/LowPowerDisplayPattern.cpp
