+++ LegacySignController/src/Patterns/DisplayPatterns/LowPowerDisplayPattern.h
