+++ LegacySignController/src/Patterns/DisplayPatterns/RadialDisplayPattern.cpp
