+++ LegacySignController/src/Patterns/DisplayPatterns/RandomDisplayPattern.cpp
