+++ LegacySignController/src/Patterns/DisplayPatterns/RandomDisplayPattern.h
