+++ LegacySignController/src/Patterns/DisplayPatterns/RotationDisplayPattern.cpp
