+++ LegacySignController/src/Patterns/DisplayPatterns/RotationDisplayPattern.h
