+++ LegacySignController/src/Patterns/DisplayPatterns/SimpleShiftDisplayPattern.cpp
