+++ LegacySignController/src/Patterns/DisplayPatterns/SimpleShiftDisplayPattern.h
