+++ LegacySignController/src/Patterns/DisplayPatterns/SolidDisplayPattern.cpp
