+++ LegacySignController/src/Patterns/DisplayPatterns/SolidDisplayPattern.h
